+++ PSModule/test.ps1
@@ -6,7 +6,7 @@ $doc = Get-M365Doc -Components Intune
 #$doc | Write-M365DocWord -FullDocumentationPath "c:\temp\$($doc.CreationDate.ToString("yyyyMMddHHmm"))-WPNinjas-Doc.docx"
 #$doc | Write-M365DocJson -FullDocumentationPath "c:\temp\$($doc.CreationDate.ToString("yyyyMMddHHmm"))-WPNinjas-Doc.json"
 
-$docnew = $doc | Optimize-M365Doc -UseTranslationFiles -UseCamelCase -ExcludeProperties @("id","@odata.type") 
+$docnew = $doc | Optimize-M365Doc -UseTranslationFiles -UseCamelCase -ExcludeProperties @("id","@odata.type") -ExcludeValues
 $docnew | Write-M365DocWord -FullDocumentationPath "c:\temp\$($doc.CreationDate.ToString("yyyyMMddHHmm"))-WPNinjas-Doc-Translated.docx"